backend/requirements.txt

Flask==2.3.3
Flask-CORS==4.0.0
ultralytics==8.0.200
opencv-python==4.8.1.78
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
